index: Fix support for logger.wrote('...')
Just ran into very strange behaviour when attempting to
`logger.type('wrote')` and `logger.wrote('stuff')`.  This
patch prefixes a private variable `Logger#wrote` with an
underscore (`Logger#_wrote`), which is *less likely* to
be used as a logging method.  It's a hack, but not sure
how else to handle this.

I'm also setting a `wrote` value after the first write,
which the prior implementation never seemed to do?